refactor: mark deps dir as private
