+++ collection/largest.test.ts
@@ -1,4 +1,4 @@
-import { assertEquals, describe, it } from "../deps/testing.ts";
+import { assertEquals, describe, it } from "../_deps/testing.ts";
 import { largest } from "./largest.ts";
 
 describe("largest", () => {
